Fix git-clone-review function when there's no PS number
When the patchset number doesn't exist yet, we might be want
to clone a given repo passing a topic which results in a failure
when git review -d is invoked.
This patch allows to continue the function execution even when
git review -d fails, checking out on the specified topic.

Signed-off-by: fmount <francesco.pan001@gmail.com>